Fixed PAN lognormalization rules.  During migration the prefix was readded to the rule body thereby misplacing the event tags (between rule= and :) further into the template leading to non-matches.
Also added an additional commented out prefix= section for loggers that dont parse the initial timestamp as part of the MSG but rather place it in the program field. Syslog-ng does this when used with the sagan output template.

And last added an empty prefix at the end of the rule section to revert the parser to non-prefixed match before proceeding.