normalization.rulebase: user-defined type for IPv4 and IPv6
This adds IPv6 support to some existing IPv4-only rules and
removes some duplication.   ::ffff:1.2.3.4 is also recognised
everywhere automatically as IPv4.

NOTE: It was necessary to enable "version=2" at the top of the file to
enable the user-defined types feature.  v2 rulebase is *more or less*
compatible with v1, but there is a risk of some breakage if some
rarely-used features of the v1 ruleset are in use.  I can't see any.

Also includes some ssh normalization fixes/additions:
- Accepted publickey ... was missing %-:rest% at the end
- Invalid user ... port %src-port:number%
- Disconnected from user ...
- Connection closed by invalid user ...

Fixes #29